+++ guillotina/db/db.py
@@ -11,6 +11,7 @@ class Root(Folder):
     __immutable_cache__ = True
     __db_id__ = None
     type_name = 'GuillotinaDBRoot'
+    migration_version = '1.0.0'  # base version
 
     def __init__(self, db_id):
         super().__init__()