+++ guillotina/db/transaction_manager.py
@@ -137,6 +137,8 @@ class TransactionManager:
                         raise
             except Exception:
                 # failsafe terminate to make sure connection is cleaned
+                if txn._db_conn is not None:
+                    raise
                 try:
                     await self._storage.terminate(txn._db_conn)
                 except asyncpg.exceptions.InterfaceError as ex: