+++ guillotina/interfaces/migration.py
@@ -1,0 +1,5 @@
+from zope.interface import Interface
+
+
+class IMigration(Interface):
+    pass