+++ guillotina/tests/test_catalog.py
@@ -14,6 +14,7 @@ from guillotina.interfaces import ICatalogDataAdapter
 from guillotina.interfaces import ICatalogUtility
 from guillotina.interfaces import ISecurityInfo
 from guillotina.tests import utils as test_utils
+from guillotina.db import TRASHED_ID
 
 
 def test_indexed_fields(dummy_guillotina, loop):
@@ -157,15 +158,15 @@ async def test_query_stored_json(container_requester):
         conn = requester.db.storage.read_conn
         result = await conn.fetch('''
 select json from {0}
-where json->>'type_name' = 'Item'
+where json->>'type_name' = 'Item' AND parent_id != '{1}'
 order by json->>'id'
-'''.format(requester.db.storage._objects_table_name))
+'''.format(requester.db.storage._objects_table_name, TRASHED_ID))
         assert len(result) == 2
         assert json.loads(result[0]['json'])['id'] == 'item1'
         assert json.loads(result[1]['json'])['id'] == 'item2'
 
         result = await conn.fetch('''
 select json from {0}
-where json->>'id' = 'item1'
-'''.format(requester.db.storage._objects_table_name))
+where json->>'id' = 'item1' AND parent_id != '{1}'
+'''.format(requester.db.storage._objects_table_name, TRASHED_ID))
         assert len(result) == 1